+++ lib/puppet/provider/oneview_server_profile_template/c7000.rb
@@ -25,9 +25,7 @@ Puppet::Type.type(:oneview_server_profile_template).provide :c7000, parent: Pupp
     prepare_environment
     empty_data_check
     connections_parse if @data['connections']
-    puppet_resource = @resource_type.new(@client, @data)
-    return false unless puppet_resource.retrieve!
-    puppet_resource.like?(@data)
+    !@resource_type.find_by(@client, @data).empty?
   end
 
   # Creates a new server profile based on the current template
@@ -47,6 +45,13 @@ Puppet::Type.type(:oneview_server_profile_template).provide :c7000, parent: Pupp
     Puppet.notice("\n\nServer Profile Template Transformation\n")
     parameters = @data.delete('queryParameters') || {}
     pretty get_single_resource_instance.get_transformation(@client, parameters)
+    true
+  end
+
+  def get_available_networks
+    Puppet.notice("\n\nServer Profile Template Available Networks\n")
+    parameters = @data.delete('queryParameters') || {}
+    pretty get_single_resource_instance.get_available_networks(@client, parameters)
     true
   end
 end